Bug 662851 - Make EnsureWritableCapacity take uint32_t instead of uint64_t since all callers pass uint32_t (r=stejohns,sr-pending=ed)

+++ core/ByteArrayGlue.cpp
@@ -119,7 +119,7 @@ namespace avmplus
         SetCopyOnWriteOwner(owner);
     }
         
-    void FASTCALL ByteArray::Grower::EnsureWritableCapacity(uint64_t minimumCapacity)
+    void FASTCALL ByteArray::Grower::EnsureWritableCapacity(uint32_t minimumCapacity)
     {
         if (minimumCapacity > (MMgc::GCHeap::kMaxObjectSize - MMgc::GCHeap::kBlockSize*2))
             m_owner->ThrowMemoryError();
